Change deltaTime to ms and added tick manager

+++ src/main/java/com/terminalvelocitycabbage/engine/client/renderer/scenes/Scene.java
@@ -15,7 +15,7 @@ public abstract class Scene {
 
 	public abstract void init();
 
-	public abstract void update(long deltaTime);
+	public abstract void update(float deltaTime);
 
 	public abstract void destroy();
